pyproject.toml
[project]
name = "apollo7"
version = "0.1.0"
description = "Photo-to-3D data sculpture pipeline"
requires-python = ">=3.12"
dependencies = [
    "PySide6>=6.8",
    "pygfx>=0.6",
    "rendercanvas>=2.6",
    "wgpu>=0.19",
    "opencv-python-headless>=4.10",
    "Pillow>=11",
    "onnxruntime-directml>=1.21",
    "numpy>=2.1",
    "scipy>=1.14",
    "extcolors",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-timeout",
    "ruff",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
timeout = 30

[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.build_meta"
